Tons of changes: -Major change in how claiming of clas and dependencies and freeing, etc works in claman and calcman interactions. -Cleaning up of transmat getting/sending -More refactoring of claman functions for clarity.  Some refactoring of pmatman along same lines, but don't think that recycling works there. -Shuffling of semi-substantial functions out of clamanager.h into condlike.cpp. -ability to throw an exception that doesn't cause program to return non-zero -addition of new tree flag, noCalcs.  Set if no holders assigned, then dirtying functions don't need to sweep dirtyness across nothing -Ability to play with fp eps at compile time (without changing source). Default with beagle is for SP=DP=FLT_EPSILON *10.  Can define DIFF_FP_EPS, then DP eps is min (DBL_EPS * 10 , 1e-12).  Or can define BEAGLE_SP_EPS and BEAGLE_DP_EPS -Default is sending transmats, unless DONT_SEND_TRANSMATS,

+++ garli/branches/beagle/src/sequencedata.h
@@ -541,7 +541,7 @@ class GeneticCode{
 		return mapIndexToCodonDisplay[index];
 		}
 
-	int NumStates() const {return mapIndexToCodonDisplay.size();}
+	unsigned NumStates() const {return (unsigned) mapIndexToCodonDisplay.size();}
 	};
 
 class CodonData : public SequenceData {
